Remoce Chorus-Pro addon
