+++ addons/fr/facturx/facturx.go
@@ -3,6 +3,7 @@ package facturx
 
 import (
 	"github.com/invopop/gobl/addons/eu/en16931"
+	"github.com/invopop/gobl/bill"
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/i18n"
 	"github.com/invopop/gobl/pkg/here"
@@ -33,12 +34,18 @@ func newAddon() *tax.AddonDef {
 				Factur-X is based on the European Norm (EN) 16931 and is mandatory for business-to-government
 				(B2G) invoices in France. This addon provides the necessary structures and validations to
 				ensure compliance with the Factur-X specifications.
-				
-				Currently this is just a placeholder addon as the EN16931 addon will provide
-				all validation requirements.
 
 				For more information on Factur-X, visit [fnfe-mpe.org](https://fnfe-mpe.org/factur-x/factur-x_en/#).
 			`),
 		},
+		Validator: validate,
 	}
+}
+
+func validate(doc any) error {
+	switch obj := doc.(type) {
+	case *bill.Invoice:
+		return validateInvoice(obj)
+	}
+	return nil
 }
