+++ data/schemas/bill/order.json
@@ -124,10 +124,6 @@
                 "title": "EN 16931-1:2017"
               },
               {
-                "const": "fr-choruspro-v1",
-                "title": "French Chorus-Pro v1"
-              },
-              {
                 "const": "fr-facturx-v1",
                 "title": "French Factur-X v1"
               },
