+++ examples/fr/invoice-fr-fr-cpp.json
@@ -1,6 +1,5 @@
 {
     "$schema": "https://gobl.org/draft-0/bill/invoice",
-    "$addons": ["fr-choruspro-v1"],
     "type": "standard",
     "series": "SAMPLE",
     "uuid": "01961f64-4934-7715-8abc-42be69de5ca3",
@@ -13,6 +12,12 @@
             "country": "FR",
             "code": "44732829320"
         },
+        "identities": [
+				{
+					"type": "SIREN",
+					"code": "732829320"
+				}
+			],
         "addresses": [
             {
                 "num": "42",
